_performance-golf/pg-creative-os/orion-chief-of-staff/_ops/daily-briefing/requirements.txt
# Exa Daily Briefing — Python Dependencies
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
slack-sdk>=3.30.0
requests>=2.31.0
pyyaml>=6.0
